+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+
+## v1.0.0-beta.9
 ### Changed
 - cedar-chart component lazy-loads cedar dependencies by default
 ### Added